adding users in sidebar
